+++ tests/test_response.py
@@ -104,7 +104,7 @@ def test_success_response_null_result():
 
 
 def test_error_response():
-    response = ErrorResponse("foo", id=1, code=-1, debug=True, http_status=200)
+    response = ErrorResponse("foo", id=1, code=-1, http_status=200)
     assert response.code == -1
     assert response.message == "foo"
     assert response.wanted == True
@@ -119,86 +119,53 @@ def test_error_response_no_id():
     # notifications under certain circumstances, such as "invalid json" and "invalid
     # json-rpc".
     assert (
-        str(ErrorResponse("foo", id=None, code=-1, debug=True, http_status=200))
+        str(ErrorResponse("foo", id=None, code=-1, http_status=200))
         == '{"jsonrpc": "2.0", "error": {"code": -1, "message": "foo"}, "id": null}'
-    )
-
-
-def test_error_response_data_with_debug_disabled():
-    # The data is not included, because debug is False
-    assert (
-        str(
-            ErrorResponse(
-                "foo", id=None, code=-1, data="bar", debug=False, http_status=200
-            )
-        )
-        == '{"jsonrpc": "2.0", "error": {"code": -1, "message": "foo"}, "id": null}'
-    )
-
-
-def test_error_response_data_with_debug_enabled():
-    assert (
-        str(
-            ErrorResponse(
-                "foo", id=None, code=-1, data="bar", debug=True, http_status=200
-            )
-        )
-        == '{"jsonrpc": "2.0", "error": {"code": -1, "message": "foo", "data": "bar"}, "id": null}'
     )
 
 
 def test_error_response_http_status():
-    response = ErrorResponse(
-        "foo", id=1, code=-1, http_status=status.HTTP_BAD_REQUEST, debug=False
-    )
+    response = ErrorResponse("foo", id=1, code=-1, http_status=status.HTTP_BAD_REQUEST)
     assert response.http_status == status.HTTP_BAD_REQUEST
 
 
 def test_invalid_json_response():
     assert (
-        str(InvalidJSONResponse(data="foo", debug=True))
+        str(InvalidJSONResponse(data="foo"))
         == '{"jsonrpc": "2.0", "error": {"code": -32700, "message": "Invalid JSON", "data": "foo"}, "id": null}'
     )
 
 
 def test_invalid_jsonrpc_response():
     assert (
-        str(InvalidJSONRPCResponse(data="foo", debug=True))
+        str(InvalidJSONRPCResponse(data="foo"))
         == '{"jsonrpc": "2.0", "error": {"code": -32600, "message": "Invalid JSON-RPC", "data": "foo"}, "id": null}'
     )
 
 
 def test_method_not_found_response():
     assert (
-        str(MethodNotFoundResponse(id=1, data="foo", debug=True))
+        str(MethodNotFoundResponse(id=1, data="foo"))
         == '{"jsonrpc": "2.0", "error": {"code": -32601, "message": "Method not found", "data": "foo"}, "id": 1}'
     )
 
 
 def test_invalid_params_response():
     assert (
-        str(InvalidParamsResponse(id=1, data="bar", debug=True))
+        str(InvalidParamsResponse(id=1, data="bar"))
         == '{"jsonrpc": "2.0", "error": {"code": -32602, "message": "Invalid parameters", "data": "bar"}, "id": 1}'
     )
 
 
 def test_exception_response():
     assert (
-        str(ExceptionResponse(ValueError("foo"), id=1, debug=True))
+        str(ExceptionResponse(ValueError("foo"), id=1))
         == '{"jsonrpc": "2.0", "error": {"code": -32000, "message": "Server error", "data": "ValueError: foo"}, "id": 1}'
     )
 
 
 def test_exception_response_with_id():
     assert (
-        str(ExceptionResponse(ValueError("foo"), id=1, debug=True))
+        str(ExceptionResponse(ValueError("foo"), id=1))
         == '{"jsonrpc": "2.0", "error": {"code": -32000, "message": "Server error", "data": "ValueError: foo"}, "id": 1}'
-    )
-
-
-def test_exception_response_debug_enabled():
-    response = ExceptionResponse(ValueError("There was an error"), id=1, debug=True)
-    assert (
-        str(response)
-        == '{"jsonrpc": "2.0", "error": {"code": -32000, "message": "Server error", "data": "ValueError: There was an error"}, "id": 1}'
     )